requirements.txt
Flask
Flask-CORS
pandas
SQLAlchemy
PyMySQL
scikit-learn
python-dotenv